separated components and rendered them onto the main code collab page
