+++ client/src/components/CodeCollabPage/PairedChat/index.tsx
@@ -3,14 +3,16 @@ import { useAppDispatch, useAppSelector } from '../../../hooks';
 import { useSocket } from '../../../context/Socket';
 import { clearChatMessages, receiveChatMessage } from '../../../store/chatRoom';
 import { PairedChatMessage } from '../../../interfaces/socket';
+import { IAgoraRTCClient, AgoraRTCProvider } from 'agora-rtc-react';
 import './index.css';
 
 // Define the props interface for the PairedChat component
 interface PairedChatProps {
     channelName: string;
+    agoraEngine: IAgoraRTCClient;
 }
 
-const PairedChat: React.FC<PairedChatProps> = ({ channelName }) => {
+const PairedChat: React.FC<PairedChatProps> = ({ channelName, agoraEngine }) => {
     const { socket, connectSocket, error } = useSocket();
     const messagesStore = useAppSelector((state) => state.chatRoom.messages);
     const [messages, setMessages] = useState<PairedChatMessage[]>([]);
@@ -74,48 +76,50 @@ const PairedChat: React.FC<PairedChatProps> = ({ channelName }) => {
 
     return (
         <>
-            <h1 className="chat-logs-header">Chat</h1>
-            <div id="messages-container" tabIndex={0}>
-                {messages &&
-                    messages.map((message, index) => {
-                        return (
-                            <div key={index} className="chat-message">
-                                <div className="message-user-img">
-                                    <img alt="" className="message-profile-pic" src={message.from.picUrl}></img>
-                                </div>
-                                <div className="message-details">
-                                    <div className="message-user-info">
-                                        {message.from.username}
-                                        <p className="message-time-updated">
-                                            {new Date(message.created_at).toLocaleTimeString([], {
-                                                hour: '2-digit',
-                                                minute: '2-digit',
-                                            })}
-                                        </p>
+            <div id="paired-chat-container">
+                <h1 className="chat-logs-header">Chat</h1>
+                <div id="messages-container" tabIndex={0}>
+                    {messages &&
+                        messages.map((message, index) => {
+                            return (
+                                <div key={index} className="chat-message">
+                                    <div className="message-user-img">
+                                        <img alt="" className="message-profile-pic" src={message.from.picUrl}></img>
                                     </div>
-                                    <p className="message-text">{message.message}</p>
+                                    <div className="message-details">
+                                        <div className="message-user-info">
+                                            {message.from.username}
+                                            <p className="message-time-updated">
+                                                {new Date(message.created_at).toLocaleTimeString([], {
+                                                    hour: '2-digit',
+                                                    minute: '2-digit',
+                                                })}
+                                            </p>
+                                        </div>
+                                        <p className="message-text">{message.message}</p>
+                                    </div>
                                 </div>
-                            </div>
-                        );
-                    })}
-            </div>
+                            );
+                        })}
+                </div>
 
-            <form className="message-input-form" onSubmit={sendChat}>
-                <textarea
-                    className="message-input"
-                    placeholder={`send message`}
-                    value={chatInput}
-                    onChange={(e) => setChatInput(e.target.value)}
-                ></textarea>
-                <button type="submit" id="send-message" aria-label="Send Message">
-                    <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-                        <path
-                            d="M21.88 4.73 16.2 20.65A2 2 0 0 1 14.3 22a2 2 0 0 1-1.9-1.31l-2.12-5.52 1.54-1.54 2.49-2.49a1 1 0 1 0-1.42-1.42l-2.49 2.49-1.58 1.55-5.51-2.13a2 2 0 0 1 0-3.83l15.96-5.68a2 2 0 0 1 2.61 2.61Z"
-                            fill="currentColor"
-                        />
-                    </svg>
-                </button>
-            </form>
+                <form className="message-input-form" onSubmit={sendChat}>
+                    <textarea
+                        className="message-input"
+                        placeholder={`send message`}
+                        value={chatInput}
+                        onChange={(e) => setChatInput(e.target.value)}
+                    ></textarea>
+                    <button type="submit" id="send-message" aria-label="Send Message">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
+                            <path
+                                d="M21.88 4.73 16.2 20.65A2 2 0 0 1 14.3 22a2 2 0 0 1-1.9-1.31l-2.12-5.52 1.54-1.54 2.49-2.49a1 1 0 1 0-1.42-1.42l-2.49 2.49-1.58 1.55-5.51-2.13a2 2 0 0 1 0-3.83l15.96-5.68a2 2 0 0 1 2.61 2.61Z"
+                                fill="currentColor"
+                            />
+                        </svg>
+                    </button>
+                </form>
+            </div>
         </>
     );
 };
