+++ client/src/components/CodeCollabPage/PairedContent/screenShareButton.tsx
